// src/app/core/api/model/updateResourcesDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContainerResourceSpecDto } from './containerResourceSpecDto';


export interface UpdateResourcesDto { 
    /**
     * Resource requests (guaranteed minimums)
     */
    requests?: ContainerResourceSpecDto;
    /**
     * Resource limits (hard ceilings)
     */
    limits?: ContainerResourceSpecDto;
    /**
     * Name of the container to update. Defaults to the first container if omitted.
     */
    containerName?: string;
}