// src/app/core/api/model/updateReplicasDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateReplicasDto { 
    /**
     * Desired replica count (0 = stopped)
     */
    replicas: number;
}